+++ package.js
@@ -6,7 +6,7 @@
 // Basic config
 // ===============================================
 const NAME = 'jest-html';
-const VERSION = '1.2.2';
+const VERSION = '1.3.0';
 const DESCRIPTION = 'Preview Jest snapshots right in your browser';
 const KEYWORDS = ['Jest', 'test', 'snapshot', 'serializer', 'html', 'preview'];
 
@@ -155,7 +155,6 @@ const specs = {
     chokidar: '1.6.1',
     'socket.io': '1.5.1',
     'lodash.debounce': '4.0.8',
-    slash: '1.0.0',
 
     // Express
     express: '^4.14.0',
@@ -178,7 +177,7 @@ const specs = {
     'react-addons-pure-render-mixin': '15.4.0', // required by giu as peer dep, delete when we update giu
     'react-frame-component': '0.6.6',
     'react-router': '4.0.0-alpha.5',
-    giu: '0.8.0',
+    giu: '0.9.0',
     'socket.io-client': '1.5.1',
     moment: '^2.0.0',
     'node-uuid': '1.4.7',