+++ src/client/gral/constants.js
@@ -1,0 +1,11 @@
+const UI = {
+  color: {
+    accentBg: '#ca461a',
+    accentFg: 'white',
+    textDim: '#999999',
+  },
+};
+
+export {
+  UI,
+};